pyproject.toml
[project]
name = "arivihan-agent"
version = "0.1.0"
description = "Autonomous AI tutoring agent for Indian students"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "openai>=1.50.0",
    "redis[hiredis]>=5.0.0",
    "boto3>=1.35.0",
    "sse-starlette>=2.0.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
